Groupwise application of functions over bed file

Groupwise application of functions over bed file

bed_group(data, vars, funs, group)

Arguments

data
Data.frame

vars
list of variable names. Must be supplied in the form list(a, b, d).

funs
List of functions to be applied to the region. See details for more.

group
Variable of denoting group

Value

The original data set with a column added for each fuction in funs.

Details

The region that the funs will be applied to is a matrix with a column for each sample and a row for each row.

Examples

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

bed_sample <- mutate(bed_sample, ggg = ceiling(row_number() / 10))
bed_group(bed_sample,
          vars = c("beta1", "beta2", "beta3", "beta4", "beta5"),
          funs = list(length, mean),
          group = ggg)

#> # A tibble: 1,000 x 3
#>      ggg length  mean
#>    <dbl>  <dbl> <dbl>
#>  1     1     50 0.670
#>  2     2     50 0.55 
#>  3     3     50 0.68 
#>  4     4     50 0.547
#>  5     5     50 0.599
#>  6     6     50 0.64 
#>  7     7     50 0.571
#>  8     8     50 0.395
#>  9     9     50 0.594
#> 10    10     50 0.564
#> # … with 990 more rows