Sliding application of functions over bed file

Sliding application of functions over bed file

bed_slide(data, vars, funs, .i, size)

Arguments

data
Data.frame

vars
list of variable names. Must be supplied in the form list(a, b, d).

funs
List of functions to be applied to the region. See details for more.

.i
Name of the variable denoting the location.

size
Size of the sliding window. The window will be centered around each site.

Value

The original data set with a column added for each fuction in funs.

Details

The region that the funs will be applied to is a matrix with a column for each sample and a row for each row.

Examples

bed_slide(bed_sample,
          vars = list(beta1, beta2, beta3, beta4, beta5),
          funs = list(length, mean),
          .i = start, size = 100)

#> # A tibble: 10,000 x 11
#>    chr    start    end coverage beta1 beta2 beta3 beta4 beta5 length  mean
#>    <fct>  <int>  <int>    <int> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#>  1 chr19 433511 433512        3     1  0     1     1     0.33     20 0.584
#>  2 chr19 433519 433520        3     1  0.67  0     0     0.67     25 0.614
#>  3 chr19 433531 433532        3     1  1     0.67  0     0.67     30 0.634
#>  4 chr19 433554 433555        3     1  0     1     0.33  0.33     30 0.634
#>  5 chr19 433565 433566        3     1  0.67  0.33  0.67  1        30 0.623
#>  6 chr19 433573 433574        3     1  0.67  1     1     0        30 0.661
#>  7 chr19 433608 433609        1     1  0     0     1     1        25 0.714
#>  8 chr19 433623 433624        2     1  1     0.5   1     0        25 0.727
#>  9 chr19 433631 433632        2     1  1     1     0     1        25 0.66 
#> 10 chr19 433666 433667        1     1  1     1     0     1        30 0.683
#> # … with 9,990 more rows